modularize existing rerun code
